refactor codes for re-set timeline response
